requirements.txt
absl-py==0.1.11
astor==0.6.2
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2017.7.27.1
click==6.7
enum34==1.1.6
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Sijax==0.4.1
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
gast==0.2.0
grpcio==1.10.0
gunicorn==19.7.1
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mock==2.0.0
numpy==1.14.1
opencv-contrib-python==3.3.0.9
pbr==3.1.1
Pillow==5.0.0
protobuf==3.5.2
scipy==1.0.0
Sijax==0.3.2
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
Werkzeug==0.14.1


python ./stylize_image.py --content ./content/ava.jpg --network-path ./pretrained-networks/dora-marr-network --output-path ./output/res.jpg


